Catch stale element reference exceptions
In `CollateEngine`, `__init__()` and `create_dataset()` have been refactored so that whenever a stale element reference exception is encountered, the latter will dump its dataset to file and pass control (along with current row indices and the path to the dump file) back to the former. `__init__()` then re-runs `create_dataset()` with the current row indices and dump file path, causing the dataset to be re-initialized from the dump file and previously-clicked rows to be skipped.

This refactoring prompted the creation of a new Exception subclass and the creation and revising of new docstrings and comments.

It seems to me that dumping the data to disk is not actually necessary. In my next commit, I will remove this functionality, but I wanted to include it in this commit in case I need it in the future.