Fix weird table indexing in Chrome, Edge
This fix ensures that tables are not indexed incorrectly in Chrome and Edge. This issue, which does not appear in Firefox, seems to be related to differences in how chromium-based browsers render elements in conjunction with javascript execution.

Unfortunately, this fix seems to cause an additional problem, which is that the amount of time to click on a row continuously increases. Ultimately, I think this indicates that a larger fix is needed, which may require re-thinking how exceptions are handled in general during the dataset creation process. The next commit will address this.